Updated package.json with more metadata
Most importantly, I added a main field, which lets node (Browserify)
know which file to import when the package is required. I also
added a link to the repo and details on the author.

I removed the Grunt task for publishing to npm in favor of
the files field in package.json, which determines what files get
included when the package is published to npm.

Lastly, I included the plugins folder, which previously didn't make
it into the npm bundle.